Fixed Ticket #56
git-svn-id: https://faracvs.cs.uni-magdeburg.de/svn/tthuem-FeatureIDE/trunk@160 272fd0a3-3bb5-487d-978a-b927a3a9602e

+++ plugins/de.ovgu.featureide.fm.ui/src/featureide/fm/ui/actions/OpenWithGuidslAction.java
@@ -95,7 +95,7 @@ public class OpenWithGuidslAction implements IObjectActionDelegate {
 		        fd.setFileName(guidslJar);
 		        guidslJar = fd.open();
 			}
-			if (new File(guidslJar).exists()) {
+			if (guidslJar != null && new File(guidslJar).exists()) { //guidslJar != null -> fixed ticket #56
 				if (newLocation) {
 					BufferedWriter writer = new BufferedWriter(new FileWriter(pathfile));
 					writer.write(guidslJar + "\r\n");
